Possible work todo with importing new bigbed info functions
There's an issue about parsing what columns mean in a BigBed file. When a bigbed is generated kent utils create their own AutoSQL, if none is provided, and base it on bed3,4,6,8,12 with other fields becoming undocumented. This could be a useful function to decipher what is going on in these libs.

There exists some previous code in ensembl-io we can adapt if we do choose to do it.