Support timeout, follow redirects and SSL verification
Lib curl has a lot of options that affect the way it does HTTP(S) communication. The interface is pretty solid but we lack a good way to transfer arbitery values through to the libcurl layer. Instead we have the three most important ones (in my opinion) that can be influenced through three static variables. A callback routine given to libBigWig is then called and sets these values on every call. We default to libCurl's default values.