Built a Perl lib of CURL constants
First step is to create a perl library that contains the CURL opt key to int value. Seems all the curl opts do is create a number so we can reuse that and provide it to CURL. WWW::Curl and Net::Curl do something similar. In order to make parsing of the headers easier we pass curl.h into cpp, which does some basic pre-processing and expands the macros curl uses.