Thread safe storage of variables
According to http://perldoc.perl.org/perlxs.html#Safely-Storing-Static-Data-in-XS we have information about how to store values safely in Perl using thread local variables. So we have implemented this scheme for the three variables we allow users to set. Other branches have tried to extend the code to enable Perl level config of CURL but it is not working well. We will keep with this version for now.