Support custom code and new emitters
The generator now allows us to inject custom code into the module. The idea is that a binary could pull this out and insert it back into the code during module rebuilds. We also support emitting an array and hash. This should allow us to read and write data from the module.